fix(palette): bound conversation content search
